Add files via upload
Q1. Reverse a Stack You are given a stack of integers. Your task is to 
reverse the order of the elements in the stack using only stack 
operations (push and pop) and without using any additional data 
structures. Ex. stack = [1, 2, 3, 4, 5] reverse Stack(stack) print(stack) 
Output should be [5, 4, 3, 2, 1].  

que 2 Three Sum Problem Sample Problem: Given an array of 
integers, find all unique triplets in the array which give the sum of 
zero. The solution should return the list of triplets.